+++ sdk/src/androidTest/java/ly/count/android/sdk/ModuleBaseTests.java
@@ -19,7 +19,9 @@ public class ModuleBaseTests {
         countlyStore.clear();
 
         mCountly = new Countly();
-        mCountly.init(new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting());
+        CountlyConfig config = new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true);
+        config.crashes.enableCrashReporting();
+        mCountly.init(config);
     }
 
     @After