+++ sdk/src/androidTest/java/ly/count/android/sdk/MultiThreadingTests.java
@@ -21,9 +21,13 @@ public class MultiThreadingTests {
 
         mCountly = new Countly();
         mCountly.halt();
-        mCountly.init(new CountlyConfig(TestUtils.getContext(), "appkey", "https://test.count.ly").setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting());
+        CountlyConfig config = new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true);
+        config.crashes.enableCrashReporting();
+        mCountly.init(config);
         mCountly.halt();
-        mCountly.init(new CountlyConfig(TestUtils.getContext(), "appkey", "https://test.count.ly").setDeviceId("1234").setLoggingEnabled(true).enableCrashReporting());
+        config = new CountlyConfig(TestUtils.getContext(), "appkey", "http://test.count.ly").setDeviceId("1234").setLoggingEnabled(true);
+        config.crashes.enableCrashReporting();
+        mCountly.init(config);
     }
 
     @Test