+++ sdk/src/main/java/ly/count/android/sdk/ModuleCrash.java
@@ -19,8 +19,6 @@ public class ModuleCrash extends ModuleBase {
 
     //crash filtering
     GlobalCrashFilterCallback globalCrashFilterCallback;
-    //Deprecated, will be removed in the future
-    CrashFilterCallback crashFilterCallback;
 
     boolean recordAllThreads = false;
 
@@ -40,14 +38,9 @@ public class ModuleCrash extends ModuleBase {
         L.v("[ModuleCrash] Initialising");
 
         globalCrashFilterCallback = config.crashes.globalCrashFilterCallback;
-        crashFilterCallback = config.crashFilterCallback;
-
         recordAllThreads = config.crashes.recordAllThreadsWithCrash;
-
         setCustomCrashSegmentsInternal(config.crashes.customCrashSegment);
-
         metricOverride = config.metricOverride;
-
         crashesInterface = new Crashes();
         breadcrumbHelper = new BreadcrumbHelper(config.sdkInternalLimits.maxBreadcrumbCount, L);
 
@@ -236,10 +229,6 @@ public class ModuleCrash extends ModuleBase {
         assert crashData != null;
 
         L.d("[ModuleCrash] Calling crashFilterCheck");
-
-        if (crashFilterCallback != null) {
-            return crashFilterCallback.filterCrash(crashData.getStackTrace());
-        }
 
         if (globalCrashFilterCallback == null) {
             return false;